multithread/requirements.txt
fastapi
uvicorn[standard]

# Streaming data
ssi-fc-data

# Data processing & DB
pandas
sqlalchemy
psycopg2-binary
orjson
kafka-python
aiokafka
